refactor: 수강신청인원을 전략패턴으로 수정
