+++ src/main/java/nextstep/courses/domain/UnlimitedCapacity.java
@@ -1,0 +1,10 @@
+package nextstep.courses.domain;
+
+public class UnlimitedCapacity implements StudentCapacity {
+
+    @Override
+    public boolean isApplicable(int count) {
+        return true;
+    }
+
+}
