+++ src/test/java/nextstep/courses/domain/LimitedCapacityTest.java
@@ -6,13 +6,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class StudentCapacityTest {
+class LimitedCapacityTest {
 
     @ParameterizedTest
     @ValueSource(ints = {0, -1})
     @DisplayName("수강신청 인원이 0이하일 경우 예외가 발생한다.")
     void throwExceptionWhenCapacityLessThanZero(int capacity) {
-        assertThatThrownBy(() -> new StudentCapacity(capacity))
+        assertThatThrownBy(() -> new LimitedCapacity(capacity))
             .isInstanceOf(IllegalArgumentException.class)
             .hasMessage("수강 인원은 1명 이상이어야 합니다.");
     }
